Moving Notification to dedicated Service namespace
