+++ BTCPayServer/Hosting/BTCPayServerServices.cs
@@ -46,7 +46,7 @@ using BTCPayServer.Security.Bitpay;
 using Serilog;
 using BTCPayServer.Security.GreenField;
 using BTCPayServer.Services.Labels;
-using BTCPayServer.Events.Notifications;
+using BTCPayServer.Services.Notifications;
 
 namespace BTCPayServer.Hosting
 {
